add support for swift 6 while maintaining support for 5
project that use macros and want to test
code-connect rely on swift-syntax of 6.0
the only way for them to support it if
multiple Package.swift files
are created.

I chose to make the new Swift 6 the default and
the old the 5.10 the duplicate. But you
could also decide to have it in reverse.